Refactor woql_local functionality
