+++ README.md
@@ -73,3 +73,6 @@ and(
 )
 ```
 
+## Unimplemented key functionality
+
+* As (for CSV Get handling to support previous functionality)
